Support redirecting stdout to the logging system
This allows redirecting stdout to a particular logger.  It doesn't
support the normal dynamically scoped semantics because that's not
really possible to do in a reliable way.

We could make `println` and `print` produce dynamically scoped output
when using the default stream, and also any other functions which look
up Base.stdout dynamically. But if anyone takes a reference to stdout
we'd just get surprises. So for now go with something which is reliable
and regular, but global to the process.